+++ packages/components/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add new component `<Stepper />` for showing a list of steps and progress.
 - Card component: updated default Muriel design.
 - Card component: new `description` prop.
+- Card component: new `isInactive` prop.
 
 # 2.0.0
 - Chart legend component now uses withInstanceId HOC so the ids used in several HTML elements are unique.